refactor: 코드 리팩토링
-carousel 컴포넌트 분리
